Update NSGAIIwIT to use BaseGA
