+++ package/samplers/nsgaii_with_initial_trials/__init__.py
@@ -1,7 +1,7 @@
 from ._mutations._gaussian import GaussianMutation
 from ._mutations._polynomial import PolynomialMutation
 from ._mutations._uniform import UniformMutation
-from .nsgaii_with_initial_trials import NSGAIIwITSampler
+from .sampler import NSGAIIwITSampler
 
 
 __all__ = [
